backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.0,<5.0

Django>=5.2.3,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
Pillow>=10.0.0





